Add context to manage search state
