+++ src/store/searchContext.js
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+const SearchContext = createContext({
+  search: '',
+  setSearch: () => {},
+});
+
+export default SearchContext;
